requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
pydantic
python-dotenv
python-jose[cryptography]
